chore: update phpstan version and remove phpstan-baseline

See merge request quiqqer/package-menu!36

+++ src/QUI/Menu/Independent/Factory.php
@@ -24,8 +24,7 @@ class Factory
             'data' => ''
         ]);
 
-        $lastId = QUI::getPDO()->lastInsertId();
-
+        $lastId = (int)QUI::getPDO()->lastInsertId();
         $Menu = Handler::getMenu($lastId);
 
         try {
